Add array support for multiple distributions
Implementation of a leaf layer that allows for differentent exponential
families for different RV subsets. E.g. for in input with 10 RV we may
want RVs 0-3 and 9 to be Gaussian, 4-7 to be Poisson and 8 to be
Categorical.

The main idea is to pass a dictionary of exponential family type that
maps to a list of indices (scopes).
E.g. for the above described example:
{
    NormalArray:  [0,1,2,3,9]
    PoissonArray:  [4,5,6,7]
    CategoricalArray:  [8]
}

Signed-off-by: Steven Lang <steven.lang.mz@gmail.com>